DataViewer panel refactoring
Moved data viewer into its own component.
